requirements.txt
Pillow>=10.0.0
tomli
tomli-w
pyinstaller